// src/api/back/models/ParticipantResultDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    BracketParticipantDto,
    BracketParticipantDtoFromJSON,
    BracketParticipantDtoFromJSONTyped,
    BracketParticipantDtoToJSON,
    OpponentResult,
    OpponentResultFromJSON,
    OpponentResultFromJSONTyped,
    OpponentResultToJSON,
} from './';

/**
 * 
 * @export
 * @interface ParticipantResultDto
 */
export interface ParticipantResultDto {
    /**
     * 
     * @type {OpponentResult}
     * @memberof ParticipantResultDto
     */
    result?: OpponentResult;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResultDto
     */
    id: number | null;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResultDto
     */
    position?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ParticipantResultDto
     */
    forfeit?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResultDto
     */
    score?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ParticipantResultDto
     */
    tbd?: boolean;
    /**
     * 
     * @type {BracketParticipantDto}
     * @memberof ParticipantResultDto
     */
    participant?: BracketParticipantDto;
}

export function ParticipantResultDtoFromJSON(json: any): ParticipantResultDto {
    return ParticipantResultDtoFromJSONTyped(json, false);
}

export function ParticipantResultDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ParticipantResultDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'result': !exists(json, 'result') ? undefined : OpponentResultFromJSON(json['result']),
        'id': json['id'],
        'position': !exists(json, 'position') ? undefined : json['position'],
        'forfeit': !exists(json, 'forfeit') ? undefined : json['forfeit'],
        'score': !exists(json, 'score') ? undefined : json['score'],
        'tbd': !exists(json, 'tbd') ? undefined : json['tbd'],
        'participant': !exists(json, 'participant') ? undefined : BracketParticipantDtoFromJSON(json['participant']),
    };
}

export function ParticipantResultDtoToJSON(value?: ParticipantResultDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'result': OpponentResultToJSON(value.result),
        'id': value.id,
        'position': value.position,
        'forfeit': value.forfeit,
        'score': value.score,
        'tbd': value.tbd,
        'participant': BracketParticipantDtoToJSON(value.participant),
    };
}